Capitalize filenames
