+++ YUViewLib/src/ui/Mainwindow.h
@@ -37,7 +37,7 @@
 #include <QSettings>
 
 #include <handler/UpdateHandler.h>
-#include <ui/separateWindow.h>
+#include <ui/SeparateWindow.h>
 #include <video/VideoCache.h>
 
 #include "ui_mainwindow.h"
