+++ YUViewLib/src/ui/PlaybackController.cpp
@@ -30,7 +30,7 @@
  *   along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "playbackController.h"
+#include "PlaybackController.h"
 
 #include <QSettings>
