+++ YUViewLib/src/ui/SettingsDialog.h
