+++ YUViewLib/src/ui/StatisticsStyleControl_ColorMapEditor.h
