+++ YUViewLib/src/ui/Statisticsstylecontrol.h
