+++ YUViewLib/src/ui/views/PlotModel.cpp
@@ -30,7 +30,7 @@
 *   along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "plotModel.h"
+#include "PlotModel.h"
 
 PlotModel::PlotModel()
 {
