+++ YUViewLib/src/ui/views/SplitViewWidget.h
